Refactor UI popups to separate module and update installer
Moved Tkinter popup functions from pikaraoke_start.py to a new assets/pikaraoke_ui.py module for better code organization and reuse. Updated pikaraoke_start.py to import and use the new UI module. Modified install.py to copy the new UI module during installation and commented out the desktop launcher installation.
